requirements.txt
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.2
Deprecated==1.2.14
dill==0.3.6
fastapi==0.100.0
flake8==6.0.0
flake8-bugbear==23.7.10
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-resources==6.0.0
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
limits==3.5.0
mccabe==0.7.0
packaging==23.1
pep8-naming==0.13.3
platformdirs==3.9.1
pluggy==1.2.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.11
pydantic-settings==1.99
pydantic_core==2.3.0
pyflakes==3.0.1
pylint==2.17.4
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
slowapi==0.1.8
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomlkit==0.11.8
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.0
wrapt==1.15.0
